wahoo: define the product image
Introduce the productpartition to wahoo series devices, however only define the common product partition in this tree.
Later details, such as the inclusion of system_ext partitions for taimen, is included in per-device trees.
Product partition is defined at 3.0GiB with later requirements of Android 13, causing an increase from the previous values.